backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
clerk_backend_api 
svix
mysqlclient
BeautifulSoup4
playwright
langchain
langchain-google-genai
langchain-community
aiomysql
redis 
fastapi-cache2
chromadb
google-generativeai
langchain-chroma
sentence-transformers
google-genai
cloudinary
pillow